requirements.txt
django==2.2.4
celery==4.3.0
django-simple-pagination==1.3
django-compressor==2.3
sorl-thumbnail==12.5.0
django-phonenumber-field==2.2.0
phonenumbers==8.10.4
django-storages==1.7
psycopg2-binary==2.8.3
arrow==0.14.5
requests==2.22.0
boto3==1.9.212
lxml==4.4.1
cssselect==1.1.0
xlrd==1.2.0
xlwt==1.3.0
openpyxl==2.6.3
pdfkit==0.6.1
redis==3.3.8
coverage==4.5.4
raven==6.10.0
sorl-thumbnail==12.5.0
pytest==5.1.1
pytest-django==3.5.1
codacy-coverage==1.3.11
django-haystack==2.8.1
elasticsearch==5.5.3
python-memcached==1.59
djangorestframework
djangorestframework-simplejwt
drf-yasg
Pillow